if de shinborikku rinku attara tukuranai

+++ setup.sh
@@ -17,7 +17,10 @@ git config --global alias.br 'branch'
 DOT_FILES=( .bashrc .vimrc .gvimrc .vim .screenrc )
 for file in ${DOT_FILES[@]}
 do
+if [ ! -e $HOME/$file ]
+then
     ln -s $HOME/dotfiles/$file $HOME/$file
+fi
 done
 
 #local環境設定ファイル
